+++ x/inflation/keeper/genesis_test.go
@@ -6,7 +6,7 @@ import (
 
 func (suite *KeeperTestSuite) TestInitGenesis() {
 	// check calculated epochMintProvision at genesis
-	epochMintProvision, _ := suite.app.InflationKeeper.GetEpochMintProvision(suite.ctx)
+	epochMintProvision := suite.app.InflationKeeper.GetEpochMintProvision(suite.ctx)
 	expMintProvision := sdk.MustNewDecFromStr("847602739726027397260274.000000000000000000")
 	suite.Require().Equal(expMintProvision, epochMintProvision)
 }